% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isAllUnique.R
\name{isAllUnique}
\alias{isAllUnique}
\title{Is All Unique Function}
\usage{
isAllUnique(OurString)
}
\arguments{
\item{OurString}{A String}
}
\description{
This function is used to check whether all the chars in a given string is all unique or not.
\cr You have to give the string as an argument to the function, shown as "OurString" below.
\cr If the chars are all unique, it prints "All unique" to the console.
\cr If the chars are not all unique, it prints "Duplicates Found." to the screen.
\cr
\cr Caution: The function is case sensitive.
}
\examples{
ExampleString<-"AaBCD"

isAllUnique(ExampleString)

 > "All unique"

 ------------------------------------

ExampleString2<-"ABA"

isAllUnique(ExampleString2)

> "Duplicates Found."

}
