% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isPermPalindrome.R
\name{isPermPalindrome}
\alias{isPermPalindrome}
\title{Is Perm Palindrome Function}
\usage{
isPermPalindrome(StringToBeTested)
}
\arguments{
\item{StringToBeTested}{A String}
}
\value{
boolean true / false
}
\description{
This function is used to check whether the chars in a given string is permutation of a palindrome, or not.
\cr You have to give the string as an argument to the function, shown as "StringToBeTested" below.
\cr If the string is a permutation of a character palindrome, it prints "The string is a palindrome" to the console.
\cr If the string is NOT a permutation of a character palindrome, it prints"The string is not a palindrome" to the screen.
\cr
\cr Caution: The function is case sensitive.
}
\examples{
ExampleString <- "kayak"
isPermPalindrome(ExampleString)

> "The string is a permutation of palindrome"
   Return T
 ------------------------------------

 ExampleString2 <- "yakak"
isPermPalindrome(ExampleString2)

> "The string is a permutation of palindrome"
   Return T
------------------------------------

 ExampleString3 <- "KAyak"
isPermPalindrome(ExampleString3)

> "The string is not a permutation of palindrome"
  Return F

}
