% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingNOs.R
\name{missingNOs}
\alias{missingNOs}
\title{Missing Numbers Function}
\usage{
missingNOs(ListToBeTested)
}
\arguments{
\item{list}{ListToBeTested   which is a list of integers}
}
\description{
Imagine you have a bag full of numbered tiles from 1 to 100 and someone takes some of them out.
\cr All you have to do is send a list of numbers to the function as an argument.
\cr It prints the missing numbers to the console.
}
\examples{
Set an example list to test.
 p1<-c(1:45)
 p2<-c(47:82)
 p3<-c(90:98)
 GivenList<-c(p1,p2,p3)

 missingNOs(GivenList)

 46
 83
 84
 85
 86
 87
 88
 89
 99





}
